refactor: address PR review — DTOs, restore missing configs and seed data
