+++ backend/src/main/java/com/senior/spm/controller/request/AdvisorOverrideRequest.java
@@ -1,4 +1,4 @@
-package com.senior.spm.controller.dto;
+package com.senior.spm.controller.request;
 
 import java.util.UUID;
