+++ backend/src/main/java/com/senior/spm/controller/request/SanitizationTriggerRequest.java
@@ -1,4 +1,4 @@
-package com.senior.spm.controller.dto;
+package com.senior.spm.controller.request;
 
 import lombok.Data;
