+++ backend/src/main/java/com/senior/spm/controller/response/AdvisorOverrideResponse.java
@@ -1,4 +1,4 @@
-package com.senior.spm.controller.dto;
+package com.senior.spm.controller.response;
 
 import java.util.UUID;
