+++ backend/src/main/resources/data.sql
@@ -51,6 +51,51 @@ VALUES
 		'PROFESSOR'
 	);
 
+INSERT IGNORE INTO password_reset_token (
+    id,
+    staff_id,
+    token,
+    created_at,
+    expires_at
+)
+VALUES (
+    UUID_TO_BIN('11111111-1111-1111-1111-111111111111'),
+    UUID_TO_BIN('00000000-0000-0000-0000-000000000004'),
+    'TEST_RESET_TOKEN_123',
+    NOW(),
+    DATE_ADD(NOW(), INTERVAL 1 DAY)
+);
+
+INSERT IGNORE INTO password_reset_token (
+    id,
+    staff_id,
+    token,
+    created_at,
+    expires_at
+)
+VALUES (
+    UUID_TO_BIN('22222222-2222-2222-2222-222222222222'),
+    UUID_TO_BIN('00000000-0000-0000-0000-000000000004'),
+    'gercek_expired_token',
+    DATE_SUB(NOW(), INTERVAL 2 DAY),
+    DATE_SUB(NOW(), INTERVAL 1 DAY)
+);
+
+INSERT IGNORE INTO password_reset_token (
+    id,
+    staff_id,
+    token,
+    created_at,
+    expires_at
+)
+VALUES (
+    UUID_TO_BIN('33333333-3333-3333-3333-333333333333'),
+    UUID_TO_BIN('00000000-0000-0000-0000-000000000004'),
+    'USED_TOKEN_123',
+    NOW(),
+    DATE_ADD(NOW(), INTERVAL 1 DAY)
+);
+
 -- Issue 55: Red Team system_config seed data
 INSERT IGNORE INTO system_config (config_key, config_value) VALUES ('active_term_id', '2026-SPRING');
 INSERT IGNORE INTO system_config (config_key, config_value) VALUES ('max_team_size', '4');
