+++ backend/src/test/java/com/senior/spm/controller/dto/InvitationResponseSerializationTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
+import com.senior.spm.controller.response.InvitationResponse;
 import org.junit.jupiter.api.Test;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
