+++ backend/src/test/java/com/senior/spm/service/GroupServiceTest.java
@@ -24,8 +24,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.senior.spm.controller.dto.BindToolResponse;
-import com.senior.spm.controller.dto.GroupDetailResponse;
+import com.senior.spm.controller.response.BindToolResponse;
+import com.senior.spm.controller.response.GroupDetailResponse;
 import com.senior.spm.entity.AdvisorRequest.RequestStatus;
 import com.senior.spm.entity.GroupInvitation.InvitationStatus;
 import com.senior.spm.entity.GroupMembership;
